+++ changelog.md
@@ -3,6 +3,7 @@
 - read endpoint from mcaprc file
 - set bikini as default project template
 - update to new mcap-generate version
+- Bugfix generate model attributes
 
 ## 0.0.5 - 2014-07-29
 - rename command to mcap